Bug #3381 - make rehash tool 64-bit safe
Thanks Leena Heino <Leena.Heino@uta.fi> for the patch.

+++ lib/util.c
@@ -269,7 +269,7 @@ int dir_hash_c(const char *name, int full)
 
     if (full) {
 	unsigned char *pt;
-	unsigned int n;
+	unsigned long long int n;
 	enum {
 	    DIR_X = 3,
 	    DIR_Y = 5,
